refactor: move venue tile files to /components as /atom doesn't exist in new code structure
